requirements.txt
numpy
Pillow
ruff
pre-commit
